removal of ResortSnowInfo and SkiReportManager
These classes were from the original implementation and
aren't needed with the new stuff like Report and ResortManager.

I also changed the callback functionality from the service
to return reports one at time so that the Dashboard won't be
as sluggish.

NOTE: This change removes the error handling logic of that
activity. However, we are going to be reworking that soon so
it should be okay for now.