Fix "loading forever" issue in dashboard
The ReportListAdapter display's a "loading" entry until
addReport(null) is call. However, when no resorts are
configured, we never launch the service that would call this.

This fix updates the Dashboard to always launch the Service
intent. It also adds logic to the ReportListAdapter to detect
an empty list and display a message to propmpt the user to
configure a report.

Closes #9