Ensure ResortManager keeps items alphabetized
This simplifies the ResortManager to just use an array instead
of a hashmap. The hashmap leads to inconsistent ordering of
resorts. In addition I run Array.sort on the list to keep them
sorted.

Closes #6

NOTE: The ResortListActivity can still add an item that won't
show up as sorted. This is being fixed with issue 7.