Create ReportController
Our API through the service to get reports has a few difficulties. We
tend to load reports more often than needed. We also don't handle
events like resort added/deleted well. In addition I've seen issues
in the original Android Ski Report application where the Activity
struggles to deal with the phone changing orientation (landscape/
portrait) while reports are loading. I studied the Android email
application Email.git and copied the pattern they use to manage
this issue. The result is better performance and a simpler model
to code to.

Noteable changes--
AndroidManifest.xml:
Created a new application class that will start the ReportController
any time this application is being run.

ReportController/ReportListener:
Pretty much stole this idea from the android Email.git project. This
provides a nice way to separate report management from the UI's. It
also introduces some performance improvements because reports aren't
always required to be loaded for things like:
 * new resort added
 * landscape/portrait mode change for the phone

WakeMeSkiService:
Removed the listener stuff that was used between this and the
dashboard. This is now used purely for alarm handling.

ReportListAdapter:
Connected this to the new ReportController framework.

ResortListActivity:
Call the ReportController so that changes reflect in the dashboard.

Dashboard:
Removed link to WakeMeSkiService. This now just listens to to the
ReportController to know when it should display the "loading" icon.