properly handle errors from bad reports
During the merge of the WakeMeSki and Android Ski Report we
lost logic that handles errors that might occur when loading
a report while doing refactoring work. This commit brings that
support back.

Here's how the flow works:

1) Report.java catches any exception occurring while loading a
report. Rather than continuing to load the report (which will
cause null-pointer issues) we exit with an error message set.

2) The main view which is rendered by ReportListAdapter.java now
has logic to display "ERROR!" in the snow fall column of the view.

3) If a user chooses to drill down from the main view, the
ReportActivity.java view will now display the actual error message
(which is likely just a e.getLocalizedMessage()).

Also fixed in this commit is a compiler warning in
AlarmController.java