Add support for multiple servers
Replace LOC_SERVER in HTTPUtils class with a
list of available servers in SERVER_LIST.

Use the server revision found in a server_info.php
file on the server to determine which server is
most recent and therefore should be used by the client.
This adds support for redundancy and also the ability
to update only one of the supported servers at a time while
redirecting all client traffic to the most recently updated server.