Fix NPE and fresh snow not found condition
NPE occurred when an error occurred fetching a
resort report.
Also, -1 was returned when fresh snow totals could
not be obtained for a resort.  This looked like a bug
when it could in actuality just be too early to determine
(not currently available.)  Modified to return N/A as
the string value instead.