Move ap versions out of reports add latest version
Move server version and latest server determination out of HttpUtils class
and into WakeMeSkiServer.
Create WakeMeSkiServerInfo.java with support for values in server_info.php.
Support ap.latest.version value in server info structure within the dashboard,
with recommended upgrade.  If the user decides not to upgrade, save the
preference and don't bug them again until the next upgrade is available.

Change the wording on the ap.min.supported.version check to be more clear about
why the upgrade is required in this case.