Cleanup error message and localization
Currently the error message returned from the server is not localized
and is in techy-nerd speak.  Modify the report class to
have the concept of both localized and non-localized error messages.

When an non-localized error message is returned from the server (ie
because a report is broken) tell the user we are working on it and
thank them for their patience

Also print "Not Available" in the summary page rather than
ERROR! to make it more obvious this is not actually an error with
the client ap.

Also fixed some centering issues on the report.xml layout and
removed unused headers on the Report activity (by removing titles)
when an error condition occurs.