V2 with preferences and alarm intent fix
Noticed a problem with the preferences page
where an out-of-date alarm could be set due to
cached values in member variables not matching
setOnPreferenceChange.

Noticed in one case the dashboard did not open
after alarm fired, instead went back to preferences
page.  I believe this was due to ACTION_MAIN in
Intent.

Added detail to string requesting the user to
enable at least one wakeup day to enable the alarm.

Bumped version to 2.0