Add resort with wakeup off fix multiple add
Fixed bug with add of multiple of the same resorts
to the list - two entries of the same resort would
show up in the dashboard view.

Also went back to wakeup enabled=false by default
when a resort is added to the list.  It's probably
better to explicitly set a wakeup for a resort if
you want one.