Fix bug in alarm calculator
Bug would place the alarm on the wrong day if the alarm should
have been placed on the current day and the minute value when the
alarm was set was greater than the minute value set for the alarm.

Needed to check for equal hours before comparing minutes value.

Also added comments