Fix alarm failure to fire (issue 20)
See github issue 20.
Made several changes to ensure the WakeMeSkiService can stay alive
in low memory conditions and guarantee we will check resort status
and fire alarm.  Most important change is change to ReportController
to only set background priority on the process when running in
conjunction with a UI thread.  Also added setForeground() to run the
service thread in foreground with notification (cross-platform support
for 1.5-2.2) and ability to set service start state to either
START_NOT_STICKY or START_REDELIVER_INTENT on 2.0 and above
 with onStartCommand()