enable support in core objects for snow alerts
This adds the logic to the Report and Weather objects to detect
when the upcoming weather report is predicting snow.

We also add a new field to Weather for the exact start time of
the report. This will be used in higher level code to make sure
we don't alert a user more than once for a given snow alert