implement a storage backend for snow alerts
This changeset provides a sqlite schema to manage the snow alerts
we pull in from our weather information in a report. This is
currently only optimized and tested for US based resorts. The base
code should, however, work for eu_report.php in the future.

= How It Works
A snow alert is a given weather description at a certain time for
a resort. A resort can have more than one alert at any time. NOTE:
weather.gov's smallest weather report increment is 6 hours. With
these constraints the following DB setup has been defined:

 TABLE: resorts:
   _id (primary key, an integer)
     - this isn't technically needed. However, we will use a
       ResourceCursorTreeAdapter for higher-level views. This
       class requires its tables to have a column by this name.
   label (text) The label along with the url are enough to create
   url (text)
     - The label along with the url are enough to create
       a Location object in code

 TABLE: alerts:
   _id (primary key, an integer)
     - this isn't technically needed(the time and resort columns
       should suffice). However, we will use a
       ResourceCursorTreeAdapter for higher-level views. This
       class requires its tables to have a column by this name.
   time (integer)
     - time in seconds since Epoch
   desc (text)
     - the actual weather report description
   acked (integer)
     - weather or not the user has been alerted about this
   resort (integer, foreign key)
     - links the alert to the resort its for

= Nuances:
Most of this is relatively straightforward android/sqlite stuff.
One interesting item is the "view_resorts" view that was created. As
we prune old alerts from the database, there's a good chance we'll
have a resort, but not alerts for it. Our higher-level code using
the ResourceCursorTreeAdapter needs a way to query for just resorts
that have alerts. The best way to do this was with a custom query. We
also had to rename the columns returned by the query for two reasons:
 1) must have a column named _id
 2) there is documented issue in the sqlite Cursor implementation that
    prevents access to columns in the <table>.<column> format.