UI support for snow alerts
This provides an actual UI and alerting mechanism based off the
backend added in the previous commit.

Code Flows:
1) The WakeMeSkiDashboard now has a menu item to launch the
AlertsActivity intent. This displays all snow alerts.

2) The ReportController now notifies the AlertManager when its done
loading reports. This allows the AlertManager to then query its
data and see if there are any un-acknowledged alerts. If there are
any, it will create (or update an existing) notification.

Noteable:

To keep things easy for the user, we consider the act of viewing the
AlertsActivity to be the acknowledgement of *all* snow alerts. This
could be made more fine grained, but the useabilty might be bad unless
handled in a clever way.

The AlertsActivity also takes the responsibiltiy for removing old alerts.
We consider old/stale alerts to be over 6 hours old (weather.gov provides
them every 6 hours so this should work correctly in the US)