Proposed notification preferences layout
To support the notification features Andy is adding, refactor
the preferences page to group wakeup and notification features
in seperate preferences groups.  Add support for a snow settings
preferences specific to notifications and an enable for notifications.

This patch does not yet tie-in preference values for notification
into the notification logic, just used as an example of how we
could re-format to include notification preferences.