Connect alert actions with preferences
Cleanup naming of alert vs notification features.  Make
alerts work seperately from notifications.  Notifications
can be disabled, alerts and the alert dialog cannot.  The
snow setting for alerts controls both whether the forecast
snow total is listed in the alert menu and whether the
alert is used to notify the user through a notification.  The
notification enabled turns on/of the user notification but
does not impact the displaying of alerts in AlertsActivity.

Two anomalies with the current check-in
1) We don't actually detect a 1 inch snow total since
it's displayed as "accumulation of one inch" instead of "accumulation 0 to 1"
It would probably be better to remove that as an option.  In reality
nobody will really care about the difference between one and two inches of
snow and nobody will probably actually want an alert on a single inch of snow?

2) It would probably be better to save the snow total in the database
rather than checking totals in the Weather class only.  This would allow
us to filter alerts already in the database rather than filtering
only before they are placed in the database.  The way it works now, changing
snow thresholds after an alert was generated will not remove it from
the AlertsActivity.