Add snow alert background polling through service
This change breaks WakeMeSkiService into WakeMeSkiAlertService (base)
and WakeMeSkiWakeupService.  WakeMeSkiAlertService manages report
loading from the service and alert notifications.  WakeMeSkiWakeupService
extends this functionality to support wakeup alarms.

Action ALERT_CHECK is handled by the alert service only.  This is fired
about once per hour (but up to once every two hours) using setInexactRepeating
and INTERVAL_HOUR with AlarmManager.  When fired, the resort reports are
obtained for all reports and used to update alerts and also place a notification
in the status bar if configured.

Action ALARM_SCHEDULE is handled by both the alert and wakeup service (by sending
intent to the wakeup service.)  This is used to schedule the wakeup alarm if
set and the alert check if notifications are enabled.

All other actions are handled by the WakeupService only (snooze, wake checks)

ALERT_CHECK is scheduled through AlertPollingController on application start
or within the preference manager when NOTIFY_ENABLE_PREF_KEY is set (the default.)
 ALERT_CHECK is disabled when NOTIFY_ENABLE_PREF_KEY is cleared since there's no
reason to do alert checks in the background if we aren't notifying the user.