Ensure Cursor/DB from AlertManager get closed
The android emulator did not complain. However, phones connected
to ADB will throw exceptions complaining about the database and
cursors not getting closed. This solves all the stack traces I
was hitting.

NOTE: WakeMeSkiAlertService did not to access the alertmanager
because the reportcontroller already handles this for it

This fixes issue 24.