Fix multiple report load on ap startup
After adding the alert service, multiple report load actions were
queue'd in ReportController on application start, resulting in
a load-clear-load sequence within the ReportActivity that was
annoying.

Modified ReportController to contain knowledge about cache invalid
conditions and whether the report list kept in the controller was
invalid due to no load ever being completed or > 1 hour since the
last load completed.  The > 1 hour check is there just because it
would theoretically be possible for the application to hang around
for a long time without being removed on some device with lots of
extra memory and not many applications running.
Modified the listner interface to remove the getReportList method
and instead always use the listner interface to update reports.
In the case reports exist in the list on the call to
addListnerAndUpdateReports() they will be reported to this listner
before adding to the mListners list.