Fix dashboard refresh corner conditions
Previous checkin of ReportController commit caused problems with
add/remove of resorts handling and refresh.  In reviewing code,
several race conditions around report list logic in ReportListAdapter
were identified.

This commit refactors ReportListAdapter and ReportController to work
more like an android ContentProvider class.  Rather than attempting
to keep a list of resorts current in ReportListAdapter through
listner methods, the ReportListAdapter listens for an onUpdated and
then queries ReportController to get a new sorted list of report
data.  This simplifies the logic in ReportListAdapter and ReportController
and gets rid of corner case synch issues.

During modification, I also noticed some problems with the SQL statements
in AlertManager.  In several cases close() was called on a member
statement, resulting in an exception the next time this member statement
was attempted to be used.  Modified to move all SQL statements into
functions where they were actually used.  I needed this modification to
test the change in this commit, hence I checked in on the same commit.

As part of this change, moved the "add resort" feature of the dashboard
when no resorts are configured to a button in the WakeMeSkiDashboard class
instead of the ReportListAdapter.

One casualty of this commit is the "Loading..." message on report load.
We can add this back in if we thik it's useful, however I think it would
be better to do so in the WakeMeSkiDashboard class than in the list
adapter class