Alert removal support and updated notify icon
If you accidentally add a resort to the list of configured resorts,
then change the resort list any old alerts for previous resorts
continue to be found in the alert activity until they expire.
This change removes alerts from the alert activity when a corresponding
resort is removed from the list of configured resorts.

This is also useful for testing because now it's easier to generate
a new alert.

Also - update the snow alert icon to contain a background.  That way
it's more easy to distinguish in the notification bar.