Add notification config button to alerts activity
We'd like to make sure snow alert notifications are visible to
all users on the next ap deploy.  We don't want people to uninstall
because they are too annoying.  Add a "configure notifications"
button to the top of the alerts preference page to allow the
user to disable or tweak notification settings easily.

Also re-word notification settings to make it more obvious these
are just status bar notifications and not alarms.