don't make assumptions on number of matches
The current alert pattern matcher assumed 2 matches. This
is the case, but I almost ran into an issue in Whistler that
would have only given one pattern match. This changes the code
to pick the last group match which can probably be assummed
to be the highest in a range (ie the 10 in 8-10).